Add unit tests for user account management endpoints
- Implement tests for listing user accounts with various filters including review status and registration status.
- Add tests for moving user accounts to different products, ensuring proper error handling for unknown products.
- Create tests for previewing approval and rejection notifications for pre-registered users, including edge cases for non-existent users.
- Implement search functionality tests for user accounts, ensuring access control based on user roles and verifying correct responses for registered and pre-registered users.
