+++ .github/instructions/python-tests.instructions.md
@@ -26,8 +26,10 @@ This is a multi-project monorepo with two main groups of projects in the folders
 ### File size
 
 - If a test file exceeds 1000 lines, split it into multiple files using descriptive name suffixes (e.g. `test_users_accounts_rest_registration.py` → `test_users_accounts_rest_registration_create.py`, `test_users_accounts_rest_registration_delete.py`, etc.).
-- If the split files share fixtures, move shared fixtures to a `conftest.py`. If the number of files warrants it, group them into a subfolder with its own `conftest.py` (e.g. `test_users_accounts_rest_registration/test_*.py` + `test_users_accounts_rest_registration/conftest.py`).
-```
+- If the split files share fixtures, move shared fixtures to a `conftest.py`. If the number of files warrants it, group them into a subfolder with its own `conftest.py`.
+- **Test file names must be unique across the entire project test tree**, regardless of which subfolder they live in. Generic names like `test_list.py`, `test_search.py`, or `test_create.py` are forbidden since they can easily collide with unrelated test files elsewhere. Always keep the full original filename as a prefix, even inside a dedicated subfolder:
+  - ✅ `test_users_accounts_rest_registration/test_users_accounts_rest_registration_search.py`
+  - ❌ `test_users_accounts_rest_registration/test_search.py`
 
 ## How to Run Tests
